refactor: split microbit_more.js into menus, modern, and legacy files
Extract menu constants to microbit-more-menus.js, modern API handlers
to microbit-more-modern.js, and legacy API handlers to
microbit-more-legacy.js. The entry point microbit_more.js now delegates
to these modules.

Refs #357
